requirements.txt
openai>=1.0.0
anthropic>=0.34.0
requests>=2.31.0
PyYAML>=6.0.0
